autogen_a2a_kit/requirements.txt
# AutoGen + A2A Kit - Developer Dependencies
# AutoGen은 setup.bat/sh에서 소스로 설치됨 (editable mode)

# Environment variables
python-dotenv>=1.0.0

# A2A Protocol - HTTP clients
requests>=2.31.0
httpx>=0.27.0

# Google ADK - A2A Server
google-adk>=0.2.0

# ASGI Server
uvicorn>=0.27.0

# AutoGen Studio (UI)
autogenstudio>=0.4.0

# ============================================
# AutoGen 소스 수정 안 할 경우 (일반 사용자용)
# setup.bat 대신 이것만 사용:
#   pip install -r requirements.txt
# ============================================
# autogen-agentchat>=0.4.2
# autogen-ext[openai]>=0.4.2
